User_Roles Completed

+++ app/controllers/wikis_controller.rb
@@ -1,19 +1,23 @@
 class WikisController < ApplicationController
   def index
     @wikis = Wiki.all
+    authorize @wikis
   end
 
   def show
     @wiki =Wiki.find(params[:id])
+    authorize @wiki
   end
 
   def new
     @wiki = Wiki.new
+    authorize @wiki
   end
 
   def create
      @wiki = Wiki.new(params.require(:wiki).permit(:title, :body))
      #raise # this will short-circuit the method
+     authorize @wiki
      if @wiki.save
        flash[:notice] = "Wiki was saved."
        redirect_to @wiki
@@ -21,7 +25,7 @@ class WikisController < ApplicationController
        flash[:error] = "There was an error saving the wiki. Please try again."
        render :new
      end
-   end
+  end
 
   def edit
     @wiki = Wiki.find(params[:id])
@@ -29,6 +33,8 @@ class WikisController < ApplicationController
 
   def update
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
+
     if @wiki.update_attributes(params.require(:wiki).permit(:title, :body))
       flash[:notice] = "Wiki was updated."
       redirect_to @wiki
@@ -36,10 +42,12 @@ class WikisController < ApplicationController
       flash[:error] = "There was an error saving the wiki. Please try again."
       render :edit
      end
-   end
+  end
 
   def destroy
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
+  
     if @wiki.destroy
       flash[:notice] = "Wiki was sucessfully deleted."
       redirect_to @wiki
